refactor: Moved Ticker out of global state
- Updated card header
