+++ src/state/types.ts
@@ -204,79 +204,6 @@ export interface PredictionsState {
   }
 }
 
-// Ticker
-
-// Only support specific token pairs until we find a method to validate token pairs
-export enum TokenPair {
-  BNBUSDT = 'bnbusdt',
-}
-
-/**
- * @see https://binance-docs.github.io/apidocs/spot/en/#individual-symbol-ticker-streams
- */
-export interface StreamData {
-  e: string
-  E: number
-  s: string
-  p: string
-  P: string
-  w: string
-  x: string
-  c: string
-  Q: string
-  b: string
-  B: string
-  a: string
-  A: string
-  o: string
-  h: string
-  l: string
-  v: string
-  q: string
-  O: number
-  C: number
-  F: number
-  L: number
-  n: number
-}
-
-export interface TickerStream {
-  eventType: string
-  eventTime: number
-  symbol: string
-  priceChange: number
-  priceChangePercent: number
-  weightAveragePrice: number
-  firstTrade: number
-  lastPrice: number
-  lastQuantity: number
-  bestBidPrice: number
-  bestBidQuantity: number
-  bestAskPrice: number
-  bestAskQuantity: number
-  openPrice: number
-  highPrice: number
-  lowPrice: number
-  totalTradedBaseAssetVolume: number
-  totalTradedQuoteAssetVolume: number
-  statisticsOpenTime: number
-  statisticsCloseTime: number
-  firstTradeId: number
-  lastTradeId: number
-  totalNumberOfTrades: number
-}
-
-export interface TickerData {
-  isConnected: boolean
-  data?: TickerStream
-}
-
-export interface TickerState {
-  data: {
-    [key in TokenPair]: TickerData
-  }
-}
-
 // Global state
 
 export interface State {
@@ -288,6 +215,5 @@ export interface State {
   predictions: PredictionsState
   profile: ProfileState
   teams: TeamsState
-  ticker: TickerState
   toasts: ToastsState
 }
