+++ src/views/Predictions/components/RoundCard/LiveRoundCard.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { CardBody, Flex, PlayCircleOutlineIcon, Progress, Text } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
 import { Round, Position } from 'state/types'
-import { useBnbUsdtTicker } from 'state/hooks'
+import { useBnbUsdtTicker } from 'hooks/ticker'
 import { formatBnbFromBigNumber, formatUsd, getBubbleGumBackground } from '../../helpers'
 import MultiplierArrow from './MultiplierArrow'
 import Card from './Card'
