+++ src/views/Predictions/components/RoundCard/SoonRoundCard.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { CardBody, Text, WaitIcon } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
 import { Position, Round } from 'state/types'
-import { padTime } from '../../helpers'
+import { formatRoundTime, padTime } from '../../helpers'
 import useBlockCountdown from '../../hooks/useGetBlockCountdown'
 import MultiplierArrow from './MultiplierArrow'
 import Card from './Card'
@@ -15,10 +15,8 @@ interface SoonRoundCardProps {
 
 const SoonRoundCard: React.FC<SoonRoundCardProps> = ({ round }) => {
   const TranslateString = useI18n()
-  const { minutes, seconds } = useBlockCountdown(round.startBlock)
-  const padMinute = padTime(minutes >= 0 ? minutes : 0)
-  const padSecond = padTime(seconds >= 0 ? seconds : 0)
-  const countdown = minutes > 0 || seconds > 0 ? `~${padMinute}:${padSecond}` : TranslateString(999, 'Soon')
+  const seconds = useBlockCountdown(round.startBlock)
+  const countdown = formatRoundTime(seconds)
 
   return (
     <Card>
