+++ src/views/Predictions/hooks/useGetBlockCountdown.ts
@@ -1,14 +1,12 @@
 import { BSC_BLOCK_TIME } from 'config'
-import { useEffect, useRef, useState } from 'react'
-import getTimePeriods from 'utils/getTimePeriods'
+import { useEffect, useState } from 'react'
 import { getWeb3NoAccount } from 'utils/web3'
 
 /**
  * Returns the difference of blocks and estimated seconds
  */
 const useBlockCountdown = (blockNumber: number) => {
-  const timer = useRef<ReturnType<typeof setTimeout>>(null)
-  const [currentBlock, setCurrentBlock] = useState(null)
+  const [baseBlock, setBaseBlock] = useState(null)
   const [secondsRemaining, setSecondsRemaining] = useState(0)
 
   // Step 1 - Fetch current block
@@ -16,29 +14,29 @@ const useBlockCountdown = (blockNumber: number) => {
     const fetchCurrentBlock = async () => {
       const web3 = getWeb3NoAccount()
       const currentBlockNumber = await web3.eth.getBlockNumber()
-      setCurrentBlock(currentBlockNumber)
+      setSecondsRemaining((blockNumber - currentBlockNumber) * BSC_BLOCK_TIME)
+      setBaseBlock(currentBlockNumber)
     }
 
     fetchCurrentBlock()
-  }, [blockNumber, setCurrentBlock])
+  }, [blockNumber, setBaseBlock])
 
-  // Step 2 - Start the countdown when we have successfully fetched the current block
+  // Step 2 - Start the timer when we have established a base block
   useEffect(() => {
-    // Only start the countdown if the block given is in the future
-    if (currentBlock !== null && blockNumber > currentBlock) {
-      setSecondsRemaining((blockNumber - currentBlock) * BSC_BLOCK_TIME)
+    let timer: ReturnType<typeof setTimeout> = null
 
-      timer.current = setInterval(() => {
+    if (baseBlock !== null && blockNumber >= baseBlock) {
+      timer = setInterval(() => {
         setSecondsRemaining((prevSecondsRemaining) => prevSecondsRemaining - 1)
       }, 1000)
     }
 
     return () => {
-      clearInterval(timer.current)
+      clearInterval(timer)
     }
-  }, [timer, blockNumber, currentBlock, setSecondsRemaining])
+  }, [baseBlock, blockNumber, setSecondsRemaining])
 
-  return getTimePeriods(secondsRemaining)
+  return secondsRemaining
 }
 
 export default useBlockCountdown
